requirements.txt
python-dotenv
pydub
jiwer
fuzzywuzzy[speedup]
openai
librosa
soundfile
scikit-learn
fastapi
uvicorn
olefile
python-docx
PyPDF2
gethwp
python-multipart
python-json-logger
colorlog
striprtf
sentry-sdk[fastapi]
pytest
pytest-cov
pytest-mock
httpx